新增导出医生名下所有病人信息和答卷信息功能

+++ src/main/java/com/yinxt/surveyscale/dto/ExaminationPaperCommitDTO.java
@@ -23,7 +23,7 @@ public class ExaminationPaperCommitDTO {
     /**
      * 答题时间
      */
-    private int useTime;
+    private Double useTime;
     /**
      * 答案信息
      */
